Gather together zuora concurrency limit handling code
