--- membership-attribute-service/app/services/FeatureToggleService.scala
@@ -1,9 +1,0 @@
-package services
-
-import models.FeatureToggle
-
-import scala.concurrent.Future
-
-trait FeatureToggleService extends HealthCheckableService {
-  def get(featureName: String): Future[Either[String, FeatureToggle]]
-}
